feat: demo 包使用 typescript 运行
